feat: rename File() function to FileToSlice()
The File() function follows the naming of the File() function in PHP but it isn't very clear what this function actually does. This change renames the function to FileToSlice() to make it a little more obvious. It also adds an "alias" of File() to preserve backward compatability (even though I'm not using this library in many places, I want to get in that habit).