Reorganized packages and added env package.
Rearange the packages so that they are sorted into nice clean directories. Also add the env package which I needed for another project. The env package could be used to replace the GetServerPort() as it just grabs and environment variable with a fallback and GetServerPort() is doing the same but is specific to the PORT env and has no arguments. I'll leave both for now but GetServerPort() could probably just call GetEnv(). GetEnv() needs some tests too.